rename: 회원 관련 클래스를 모두 domain/member 디렉토리 안으로 이동
