+++ thirdAssignment/src/main/java/org/sopt/domain/member/dto/response/MemberGetAllResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.dto.response;
+package org.sopt.domain.member.dto.response;
 
 import java.util.List;
